Completed links and assets clean-up
Finishing updating all of the links and moving the proper images into the "assets" folder. Starting to rename some of the files, but have not altered many at this point. Saving at this stage because I will be making a lot of file name changes next and I wanted to ensure there was a good backup.